dependency tree - dont show empty drawers
