+++ scopes/code/ui/dependency-tree/dependency-drawer/index.ts
@@ -1,0 +1,2 @@
+export { DependencyDrawer } from './dependency-drawer';
+export type { DependencyDrawerProps } from './dependency-drawer';
