--- scopes/code/ui/dependency-tree/dependency-tree.module.scss
@@ -1,111 +1,0 @@
-$panelBg: #fafafa;
-
-.codePage {
-  height: calc(100vh - 56px);
-  width: 100%;
-}
-
-.left {
-  min-width: 200px;
-
-  //cannot assign max-width directly to right side
-  max-width: calc(100% - 200px);
-
-  display: flex;
-  // overflow-y: auto;
-}
-
-.right {
-  position: relative;
-  background: $panelBg;
-  overflow-y: auto;
-  min-width: 200px;
-}
-
-.splitter {
-  position: relative;
-  border-right: 1px solid var(--border-color, #ededed);
-}
-
-.collapser {
-  right: 0;
-}
-
-.codeView {
-  padding: 24px 40px;
-  width: calc(100% - 80px);
-  overflow: hidden;
-
-  .codeSnippetWrapper {
-    width: 100%;
-    max-width: 100%;
-    .codeSnippet {
-      display: block;
-      overflow: auto;
-      height: calc(100vh - 200px);
-    }
-  }
-}
-
-.img {
-  width: 20px;
-  margin-right: 10px;
-}
-
-.fileName {
-  display: flex;
-  align-items: center;
-}
-
-.depNode {
-  display: flex;
-  align-items: center;
-  flex: 1;
-  min-width: 0;
-  //   overflow: hidden;
-  //   text-overflow: ellipsis;
-  //   white-space: nowrap;
-  font-size: 14px;
-  &:hover {
-    background: var(--bg-topping-highlight);
-  }
-}
-
-.dependencyDrawerContainer {
-  .dependencyDrawer {
-    > div:first-child {
-      font-size: 14px;
-      border: unset;
-      padding-left: 16px;
-    }
-  }
-
-  .dependencyLink {
-    height: 32px;
-    display: flex;
-    align-items: center;
-    // justify-content: space-between;
-    padding: 0 8px;
-    padding-left: 32px;
-    font-size: inherit;
-    font-size: var(--p-xs);
-    text-decoration: none;
-    color: inherit;
-
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-
-    box-sizing: border-box;
-    border-left: 0px solid transparent;
-    // > span {
-    // }
-    > span:first-child {
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      color: var(--base-color);
-      margin-right: 5px;
-    }
-  }
-}
